Use the "go" tool for building.
Change the package structure so it can be imported as
"github.com/jackpal/Taipei-Torrent".

Remove Makefiles and helper scripts. To build, use "go build" (works
with weekly.2012-01-15).

To clean, use "go clean" (not yet available in the current weekly).

Note that the go tool expects the source code to be in a very specific
directory. See: http://tip.golang.org/doc/code.html