Get acquaintance with unknow nodes.
If a node we never seen try to communicate with us, test if they are
reachable and if so add them to our routing table.

Later it might be a good idea to send them a get_peers instead of a
ping, so we don't waste bandwidth. But for now this is cleaner.