Global rate limiting.
Use a token bucket to limit the number of packets per second that I want
to receive. Beyond this limit all packets will be silently ignored.

The token bucket is updated only 10 times per second, so there will be a lot
of burstiness, but that shouldn't be a problem since this is for
receiving only (the network transmission already happened after all).