Add preliminary support for NAT-PMP port forwarding.
Example usage:

 -useNATPMP -gateway 192.168.1.1

You must manually specify the IP address of the default gateway.
Discovering the IP address of the default gateway is not part of the
NAT-PMP protocol.

Go doesn't currently seem to have a library for discovering the IP
address of the default gateway.

Adjusted the NAT interface to work with the NAT-PMP protocol.