Move all flags out of torrent library into main.go
It's generally better to keep flags out of packages.

Add a public struct, torrent.TorrentFlags, for specifying flags that
control a torrent session. Pass this struct around as needed.

Rework the socks proxy code. The torrent.TorrentFlags struct takes a
generic Dialer function argument, and only the main.go code knows about
the socks5 proxy library.

Fix case of -trackerlessMode flag. "Trackerless" is a word, so the "l"
should not be capitalized.