Download to SFTP; Skip initial check; Cache to ram
The purpose of this code change is to permit Taipei-Torrent to save the
downloaded torrent to an SFTP server instead of the local filesystem.
The reason that this might be useful is that some machines (i.e. cheap
VPS's) don't have enough harddrive space to download large torrents, and
often can't be modified to include the FUSE kernel extensions necessary
for mounting SFTP shares. In order to make this feature relatively
efficient, it's useful to skip the initial hash check on load, as well
as to cache some of the data locally (otherwise you end up transferring
the data multiple times; once to write to file, and again each time it
checks the hash). So, a flag to prevent the initial hash check has been
added, as well as an interface for caches and a slightly trivial cache
implementation using ram.