Improve RamCache, and add HdCache.
Improve RamCache by changing the chunk size to be equal to the torrent's
piece size (seems obvious in retrospect...), as well as *trying* to
commit to base storage only pieces that have passed the hash check (if
we run out of cache space, then may have to write unproven data to base
storage).

Also, I've added HdCache, which does more or less the same as RamCache
but using the OS's temporary storage folder.

As well as few other tweaks here and there.