Fix build in windows; organize a bit.
Seeding notification broke build on windows. "syscall" is a
system-dependent package, and in windows "syscall.SIGCHLD" is undefined.
Additionally, I moved the feature to its own .go file, and did a bit of
cleanup as well. Now works in both windows and linux (at least in my
machines. :-) )

Also did a go fmt.