Multiple torrents support.
Near as I can tell, and after some testing, the main thing preventing
multiple torrents from working was that each torrentSession would create
it's own DHT node. This commit moves that out to RunTorrents(). I've run
this a few times on my machine with 2 torrents concurrently, and haven't
had any problems. However, in the name of caution, I'm putting this as a
separate branch until it's had time to bake.