New Feature: QuickResume
QuickResume: By saving a copy of the pieceSet to disk (
"./*-haveBitset", where '*' is the torrent's infohash as a hex-encoded
string) every time we commit a good piece to storage, we can resume
quicker if the download gets interrupted. This file tends to be small
(size in bytes = ceiling of the number of pieces / 8). Please note that
in order for QuickResume to quickly resume a torrent:
a) The previous run must have had the QuickResume flag set to true
b) The resuming run must have the InitialCheck flag set to false