Fix: issue #69 A data race reading/writing t.si.HaveTorrent
The data race mentioned in issue #69 is the result of reading
t.si.HaveTorrent in a goroutine that's been spun up to connect to a
peer. Thankfully, this read is redundant, and can be safely removed: The
boolean expression that uses it is already being used by the call to
mightAcceptPeer(), which always precedes connectToPeer().