Prevent Self Connections
In some situations (in the case I was able to replicate, a torrent with
a low number of peers & using DHT), TaipeiTorrent will try to connect to
itself as a peer. This commit detects established self-connections,
takes note of the self-address, and closes them. The list of
self-addresses is then used to prevent new outgoing self-connections.
Thread-safe access to the OurAddresses map is not a concern because it
is only written to/read by addPeerImp() & tryNewPeer(), which are always
executed by the same goroutine (the main loop in DoTorrent()).