New Flag: memoryPerTorrent
One of the main sources of memory use is the actively downloading pieces
("active pieces"). This new flag allows the user to control this aspect
of memory use. When not used, or set to a negative number, this flag has
no effect. Otherwise, it indicates the maximum number of MiB that each
torrent can devoted to active pieces. This may have some impact on
download speeds, since it limits the number of pieces that can be
downloaded concurrently (but not necessarily the number of peers being
downloaded from, since more than one peer can contribute to a piece).
This option is most useful in systems with low available memory and/or a
large numbers of concurrently downloading torrents. Thus, when used,
this flag also sets the garbage collector to clean up more often.