fixing error in generating random ID
Fixing error in generating random ID for struct Tracker. In  randomString function, in order to generate ID on the basis of a string "0123456789abcdef", the project should have liked to generate random char from string  "0123456789abcdef" and assigns it to b[i], but it actually assigns b[] to b[i]. It is a nil array.