fix error in downloading pieces continue
In original codes, Taipei-Torrent could not download pieces from the peer which have been set to "UNINTERESTED" by it again, because it download piece continue after it has got a "PIECE". When it set "UNINTERESTED" to its peer, it stops to download. So we should boot it again after the peer have pieces it needs again.